+++ internal/proxy/usecase/proxy_ucase.go
@@ -108,7 +108,7 @@ func (pu *ProxyUcase) SaveReqToDB(request *http.Request, scheme string) error {
 	return nil
 }
 
-func (pu *ProxyUcase) GetRequest(id int) (*models.Request, error) {
+func (pu *ProxyUcase) GetRequest(id int64) (*models.Request, error) {
 	return pu.proxyRepo.GetRequest(id)
 }
 func (pu *ProxyUcase) GetAllRequests() ([]*models.Request, error) {
